refactor: extract the ic-management-canister-types crate
