+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "ic-cdk-bindgen",
     "ic-cdk-macros",
     "ic-cdk-timers",
+    "ic-management-canister-types",
     "ic-response-codes",
     "ic0",
     "library/*",
@@ -34,6 +35,7 @@ ic0 = { path = "ic0", version = "0.24.0-alpha.1" }
 ic-cdk = { path = "ic-cdk", version = "0.18.0-alpha.1" }
 ic-cdk-timers = { path = "ic-cdk-timers", version = "0.12.0-alpha.1" }
 ic-response-codes = { path = "ic-response-codes", version = "0.1.0" }
+ic-management-canister-types = { path = "ic-management-canister-types", version = "0.1.0" }
 
 candid = "0.10.12"
 candid_parser = "0.1.4"
