+++ ic-management-canister-types/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "ic-management-canister-types"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+license.workspace = true
+description = "Types for calling the IC management canister"
+homepage = "https://docs.rs/ic-management-canister-types"
+documentation = "https://docs.rs/ic-management-canister-types"
+readme = "README.md"
+categories = ["api-bindings", "data-structures"]
+keywords = ["internet-computer", "types", "dfinity", "canister"]
+include = ["src", "Cargo.toml", "LICENSE", "README.md"]
+
+[dependencies]
+candid.workspace = true
+serde.workspace = true
+serde_bytes.workspace = true
