+++ ic-management-canister-types/README.md
@@ -1,0 +1,7 @@
+# ic-management-canister-types
+
+Types for calling [the IC management canister][1].
+
+This module is a direct translation from its Candid interface description.
+
+[1]: https://internetcomputer.org/docs/current/references/ic-interface-spec/#ic-management-canister
